Switched from SQLite to PostgreSQL and got an error with telegram_id being BigInteger, not INTEGER
To fix this, I simply rolled back the migration, changed INTEGER to BigInteger in both the migration file and the model, and ran the migration again, thus changing the field in the table

In the next step, I will build and run the bot in Docker

+++ alembic/versions/ad43149a2977_initial_migration.py
@@ -28,7 +28,7 @@ def upgrade() -> None:
     )
     op.create_table('users',
     sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('telegram_id', sa.Integer(), nullable=True),
+    sa.Column('telegram_id', sa.BigInteger(), nullable=True),
     sa.Column('username', sa.String(), nullable=True),
     sa.PrimaryKeyConstraint('id'),
     sa.UniqueConstraint('telegram_id')
